requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
yfinance>=0.2.37
hmmlearn>=0.3.0
pyyaml>=6.0.0
